requirements.txt
absl-py==2.1.0
asttokens 
astunparse==1.6.3
backcall 
certifi==2024.6.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
comm 
debugpy 
decorator 
exceptiongroup 
executing 
filelock==3.13.1
flatbuffers==24.3.25
fsspec==2024.2.0
gast==0.5.4
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.64.1
h5py==3.11.0
humanfriendly==10.0
idna==3.7
importlib-metadata 
ipykernel 
ipython 
jedi 
Jinja2==3.1.3
joblib==1.4.2
jupyter_client 
jupyter_core 
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline 
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nest-asyncio 
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.16.1
onnx-simplifier==0.4.36
onnxoptimizer==0.3.13
onnxruntime-gpu==1.18.0
opt-einsum==3.3.0
optree==0.11.0
packaging 
pandas==2.2.2
parso 
pexpect 
pickleshare 
pillow==10.2.0
platformdirs 
prompt-toolkit
protobuf==4.25.3
psutil
ptyprocess
pure-eval
Pygments
python-dateutil
pytz==2024.1
pyzmq
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
six
stack-data
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorrt==10.7.0.post1
tensorrt-cu12-bindings==10.7.0.post1
tensorrt-cu12-libs==10.7.0.post1
tensorrt_cu12==10.7.0.post1
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.6.0
torch_tensorrt==2.6.0
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
torchviz==0.0.2
tornado
tqdm==4.66.4
traitlets
triton==3.2.0
typing_extensions
tzdata==2024.1
urllib3==2.2.1
wcwidth
Werkzeug==3.0.3
wrapt==1.16.0
zipp
